requirements.txt
Package                       Version
----------------------------- -----------
asgiref                       3.8.1
certifi                       2024.8.30
charset-normalizer            3.3.2
Django                        5.1.1
django-environ                0.11.2
django-filter                 24.3
djangorestframework           3.15.2
djangorestframework-gis       1.1
djangorestframework-simplejwt 5.3.1
GDAL                          3.6.2
geopandas                     1.0.1
idna                          3.10
mkl_fft                       1.3.10
mkl_random                    1.2.7
mkl-service                   2.4.0
numpy                         1.26.4
packaging                     24.1
pandas                        2.2.3
pip                           24.2
psycopg2                      2.9.9
PyJWT                         2.9.0
pyogrio                       0.10.0
pyproj                        3.7.0
python-dateutil               2.9.0.post0
python-dotenv                 1.0.1
pytz                          2024.2
requests                      2.32.3
setuptools                    75.1.0
shapely                       2.0.6
six                           1.16.0
sqlparse                      0.5.1
tzdata                        2024.2
urllib3                       2.2.3
wheel                         0.44.0
